requirements.txt
streamlit==1.32.2
pandas
numpy
plotly
python-binance
